backend: Move utils from test directory to general utils file
